+++ frontend/src/auth/AdminLogincopy.js
@@ -12,7 +12,7 @@ const AdminLogin = ({ setToken }) => {
   const handleLogin = async (event) => {
     event.preventDefault();
     try {
-      const response = await axios.post("http://localhost:8000/api/token/", {
+      const response = await axios.post("http://192.168.1.7:8000/api/token/", {
         username: username,
         password: password,
       });
