+++ frontend/src/auth/TeacherLogin.js
@@ -8,7 +8,7 @@ const AdminLogin = ({ setToken }) => {
   const handleLogin = async (event) => {
     event.preventDefault();
     try {
-      const response = await axios.post("http://localhost:8000/api/token/", {
+      const response = await axios.post("http://192.168.1.7:8000/api/token/", {
         username: username,
         password: password,
       });
