+++ frontend/src/components/Navbar/Navbar.js
@@ -1,146 +1,135 @@
-import React, { useRef, useState, useEffect } from "react";
-import "../../styles/homepage.css";
-import NavbarText from "./NavbarText";
+import React, { useState } from "react";
 import { Link } from "react-router-dom";
+import AppBar from "@mui/material/AppBar";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import Button from "@mui/material/Button";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import { useTheme } from "@mui/material/styles";
 
 const Navbar = () => {
-  const [isCoursesDropdownOpen, setIsCoursesDropdownOpen] = useState(false);
-  const coursesDropdownRef = useRef(null); // Reference for the courses dropdown
-  const [isPortalDropdownOpen, setIsPortalDropdownOpen] = useState(false);
-  const portalDropdownRef = useRef(null); // Reference for the portal dropdown
+  const theme = useTheme();
+  const [anchorEl, setAnchorEl] = useState(null);
 
-  // Function to close the dropdowns if a click is detected outside of them
-  const handleClickOutside = (event) => {
-    if (
-      coursesDropdownRef.current &&
-      !coursesDropdownRef.current.contains(event.target)
-    ) {
-      setIsCoursesDropdownOpen(false);
-    }
-    if (
-      portalDropdownRef.current &&
-      !portalDropdownRef.current.contains(event.target)
-    ) {
-      setIsPortalDropdownOpen(false);
-    }
+  const handleMenuOpen = (event) => {
+    setAnchorEl(event.currentTarget);
   };
 
-  useEffect(() => {
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
+  const handleMenuClose = () => {
+    setAnchorEl(null);
+  };
 
   return (
-    <nav className="home-navbar" aria-label="Main navigation">
-      <div className="left-section">
-        <NavbarText>
-          <Link to="/">Home</Link>
-        </NavbarText>
-        <div className="chatbot-text-navbar">
-          <Link to="/chatbot">Chatbot</Link>
-        </div>
-        <div className="dropdown" ref={coursesDropdownRef}>
-          <button
-            onClick={() => setIsCoursesDropdownOpen(!isCoursesDropdownOpen)}
-            className="dropdown-btn"
-            aria-haspopup="true"
-            aria-expanded={isCoursesDropdownOpen}
-          >
-            <NavbarText>Courses</NavbarText>
-          </button>
-          <div
-            className={`dropdown-content ${
-              isCoursesDropdownOpen ? "show" : ""
-            }`}
-            role="menu"
-          >
-            <Link
-              to="/courses/casa"
-              tabIndex={isCoursesDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              CASA
-            </Link>
-            <Link
-              to="/courses/grade-school"
-              tabIndex={isCoursesDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              Grade School
-            </Link>
-            <Link
-              to="/courses/highschool"
-              tabIndex={isCoursesDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              High School
-            </Link>
-            <Link
-              to="/courses/sped"
-              tabIndex={isCoursesDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              SPED
-            </Link>
-            <Link
-              to="/courses/homestudy"
-              tabIndex={isCoursesDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              HomeStudy{" "}
-            </Link>
-            {/* ... */}
-          </div>
-        </div>
-        <NavbarText>
-          {" "}
-          <Link to="/about">About</Link>
-        </NavbarText>
-        {/*     <NavbarText>
-          <Link to="/enroll">Apply</Link>
-        </NavbarText>
-        <NavbarText>
-          <Link to="/apply">Enroll</Link>
-        </NavbarText> */}
-      </div>
-      <div className="right-section">
-        <div className="dropdown" ref={portalDropdownRef}>
-          <button className="dropdown">
-            <button
-              onClick={() => setIsPortalDropdownOpen(!isPortalDropdownOpen)}
-              className="dropdown-btn"
-              aria-haspopup="true"
-              aria-expanded={isPortalDropdownOpen}
-            >
-              <NavbarText>LGMS Portal</NavbarText>
-            </button>
-          </button>
-          <div
-            className={`dropdown-content ${
-              isCoursesDropdownOpen ? "show" : ""
-            }`}
-            role="menu"
-          >
-            <Link
-              to="/portal/student"
-              tabIndex={isPortalDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              Student
-            </Link>
-            <Link
-              to="/portal/teacher"
-              tabIndex={isPortalDropdownOpen ? "0" : "-1"}
-              role="menuitem"
-            >
-              Teacher
-            </Link>
-          </div>
-        </div>
-      </div>
-    </nav>
+    <AppBar position="static" sx={{ backgroundColor: theme.palette.grey[100] }}>
+      <Toolbar>
+        <Typography
+          variant="body1"
+          component={Link}
+          to="/"
+          sx={{
+            marginRight: theme.spacing(2),
+            color: theme.palette.text.primary,
+            textDecoration: "none",
+          }}
+        >
+          Home
+        </Typography>
+        <Typography
+          variant="body1"
+          component={Link}
+          to="/chatbot"
+          sx={{
+            marginRight: theme.spacing(2),
+            color: "#286e34",
+            textDecoration: "none",
+          }}
+        >
+          Chatbot
+        </Typography>
+        <Button
+          aria-controls="courses-menu"
+          aria-haspopup="true"
+          onClick={handleMenuOpen}
+          sx={{
+            marginRight: theme.spacing(2),
+            color: theme.palette.text.primary,
+            textDecoration: "none",
+          }}
+        >
+          Courses
+        </Button>
+        <Menu
+          id="courses-menu"
+          anchorEl={anchorEl}
+          keepMounted
+          open={Boolean(anchorEl)}
+          onClose={handleMenuClose}
+        >
+          <MenuItem component={Link} to="/courses" onClick={handleMenuClose}>
+            All Courses
+          </MenuItem>
+        </Menu>
+        <Typography
+          variant="body1"
+          component={Link}
+          to="/about"
+          sx={{
+            marginRight: theme.spacing(2),
+            color: theme.palette.text.primary,
+            textDecoration: "none",
+          }}
+        >
+          About
+        </Typography>
+        <Typography
+          variant="body1"
+          component={Link}
+          to="/contact"
+          sx={{
+            marginRight: theme.spacing(2),
+            color: theme.palette.text.primary,
+            textDecoration: "none",
+          }}
+        >
+          Contact Us
+        </Typography>
+        <Typography
+          variant="body1"
+          component={Link}
+          to="/enroll"
+          sx={{
+            marginRight: theme.spacing(2),
+            color: theme.palette.text.primary,
+            textDecoration: "none",
+          }}
+        >
+          Enroll
+        </Typography>
+        <Button
+          aria-controls="portal-menu"
+          aria-haspopup="true"
+          onClick={handleMenuOpen}
+          sx={{
+            color: theme.palette.text.primary,
+            textDecoration: "none",
+          }}
+        >
+          LGMS Portal
+        </Button>
+        <Menu
+          id="portal-menu"
+          anchorEl={anchorEl}
+          keepMounted
+          open={Boolean(anchorEl)}
+          onClose={handleMenuClose}
+        >
+          <MenuItem component={Link} to="/portal" onClick={handleMenuClose}>
+            Portal Home
+          </MenuItem>
+        </Menu>
+      </Toolbar>
+    </AppBar>
   );
 };
