+++ frontend/src/services/api.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import { jwtDecode } from "jwt-decode";
 
 const api = axios.create({
-  baseURL: "http://localhost:8000/api",
+  baseURL: "http://192.168.1.7:8007/api",
   headers: {
     "Content-Type": "application/json",
   },
